Fix isRunningProcess on Windows
The implementation of proc.Signal on Windows is incomplete. When an
unimplemented signal is received, EWINDOWS is returned.

Instead of relying on Signal(0) to return success if a process is still
running, rely on the fact that Signal will short-circuit with a
meaningful error if the process is not running. Thus, we can ignore
EWINDOWS, which likely means the process is still running.

Signed-off-by: Matthew Horan <mhoran@pivotal.io>